@@ -2,9 +2,9 @@
  *  finds the next time slot
  */
 
-const {norm: fmt} = require('./fmt');
+import {norm as fmt} from './fmt.js';
 
-module.exports = (cycle, s, now, utc, last) => {
+export default function(cycle, s, now, utc, last) {
   let offset;
   if (utc) {
     utc = 'UTC';
@@ -92,6 +92,6 @@ module.exports = (cycle, s, now, utc, last) => {
   }
 
   return then;
-};
+}
 
 // end of nextime.js
